% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getpbf_filuf.R
\name{gepbf_filuf}
\alias{gepbf_filuf}
\title{Download a tibble with all the IBGECODE from municipality of the desired State (UF).}
\usage{
gepbf_filuf(UF_Num)
}
\arguments{
\item{UF_Num}{Brazilian Federative Unit Code, a list of them might be requested with the owdbr::getpbf_uflist() function}
}
\value{
Returns a tibble with the list of municipalities.
}
\description{
Download a tibble with all the IBGECODE from municipality of the desired State (UF).
}
\examples{
getpbf_filuf('BA')
getpbf_filuf('RJ')
getpbf_filuf('MS')


}
