% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getpbf_uflist.R
\name{getpbf_uflist}
\alias{getpbf_uflist}
\title{All the State's UF Codes.
Returns a tibble which contains the IBGE identification code of each one of all 23 UFs (Units of the Federation) in Brazil.
Note: In alphabetical order.
 @param none}
\usage{
getpbf_uflist()
}
\value{
Returns a tibble with the states in the 'EST' column and its respective UF code in the 'UF' column.
}
\description{
All the State's UF Codes.
Returns a tibble which contains the IBGE identification code of each one of all 23 UFs (Units of the Federation) in Brazil.
Note: In alphabetical order.
 @param none
}
\references{
Source: \href{https://cidades.ibge.gov.br}{IBGE (Brazilian Institute of Geography and Statistics)}
}
