% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munlist.R
\name{munlist}
\alias{munlist}
\title{Get a tibble with a all municipalities from a State.}
\usage{
munlist(UF_Num)
}
\arguments{
\item{UF_Num}{A number given by the IBGE (Brazilian Institute for Geography and Statistics) for each state, a list of them might be requested with the owdbr::uflist() function.}
}
\value{
a tibble with the requested data divided in 11 columns, if there are more than one IBGECODE, returns all of them in the same tibble.
\describe{
  \item{table.dataReferencia}{Reference date}
  \item{table.valor}{Amount of money invested in the municipality.}
  \item{table.quantidadeBeneficiados}{Number of citizens wich are enrolled in the program in that moment}
  \item{table.municipio.codigoIBGE}{IBGE code of the municipality.}
  \item{table.municipio.nomeIBGE}{Name of the municipality.}
  \item{table.municipio.pais}{Country of the municipality.}
  \item{table.municipio.uf.sigla}{Abbreviation of name of the state in which the municipality is.}
  \item{table.municipio.uf.nome}{Full name of the state in wich the municipality is.}
  \item{table.tipo.id}{Type}
  \item{table.tipo.descricao}{Abbreviation of the program's name.}
  \item{table.tipo.descricaoDetalhada}{Full name of the program.}
  }
}
\description{
Download a tibble with all the IBGECODE from municipality of the desired State.
}
\examples{
munlist(11)
munlist(c(11, 50))

}
\author{
Joao Pedro Oliveira dos Santos, International Relations Institute, Pontifical Catholic University of Rio de Janeiro
}
