% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owdbr_fies.R
\name{owdbr_fies}
\alias{owdbr_fies}
\title{Collects data related to the Brazilian Federal Government's Educational Program FIES (Financiamento Estudantil).}
\usage{
owdbr_fies(path)
}
\description{
Collects data related to the Brazilian Federal Government's Educational Program FIES (Financiamento Estudantil).
}
