% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uflist.R
\name{uflist}
\alias{uflist}
\title{All the State's UF Codes.}
\usage{
uflist()
}
\value{
a tibble with 3 columns.
\describe{
   \item{EST}{The full name of the state}
   \item{UF}{The abbreviation of the state's name}
   \item{num}{The identification number of the state.}
}
}
\description{
Returns a tibble which contains the IBGE identification code of each one of all 23 UFs (Units of the Federation) in Brazil.
}
\examples{
uflist()

}
\references{
Source: \href{https://cidades.ibge.gov.br}{IBGE (Brazilian Institute of Geography and Statistics)}
}
\author{
Joao Pedro Oliveira dos Santos, International Relations Institute, Pontifical Catholic University of Rio de Janeiro
}
