% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getpbf_mun.R
\name{getpbf_mun}
\alias{getpbf_mun}
\title{Collects municipal level data from the Program Bolsa Familia Database.}
\usage{
getpbf_mun(IBGECODE, AAAA, MM, PAGE)
}
\arguments{
\item{IBGECODE}{IBGE Unique Identifier of the municipality which you want the data. The list of municipalities and its codes might be requested with the function (INSERT FUNCTION THAT COLLECTS MUN-LIST)}

\item{AAAA}{Year of the Data (AAAA format)}

\item{MM}{Month of the Data (MM format)}

\item{PAGE}{Request's page. Default= 1.}
}
\value{
Returns a tibble with the requested data, if there are more than one IBGECODE, returns all of them in the same tibble.
}
\description{
Collects municipal level data from the Program Bolsa Familia Database.
}
